requirements.txt
torch==2.8.0
torchvision==0.23.0
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.3.0
scikit-learn==1.5.2
streamlit==1.39.0
streamlit-drawable-canvas==0.9.3
matplotlib==3.9.2
pandas==2.2.3
tqdm==4.66.5
scipy==1.14.1
torchsummary
torchviz
tensorboard